+++ .gitignore
@@ -2,7 +2,7 @@
 # Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion, Android Studio, WebStorm and Rider
 # Reference: https://intellij-support.jetbrains.com/hc/en-us/articles/206544839
 
-# User-specific stuff
+# Participant-specific stuff
 .idea/**/workspace.xml
 .idea/**/tasks.xml
 .idea/**/usage.statistics.xml
@@ -10,7 +10,7 @@
 .idea/**/shelf
 .idea/**
 
-# AWS User-specific
+# AWS Participant-specific
 .idea/**/aws.xml
 
 # Generated files
