+++ config/packages/security.yaml
@@ -7,7 +7,7 @@ security:
         # used to reload user from session & other features (e.g. switch_user)
         app_user_provider:
             entity:
-                class: App\Entity\User
+                class: App\Entity\Participant
                 property: email
     firewalls:
         dev:
