+++ templates/participant/updateProfil.html.twig
@@ -1,0 +1,6 @@
+{% extends 'base.html.twig' %}
+
+{% block body %}
+    {{ form_start(form) }}
+    {{ form_end(form) }}
+{% endblock %}
